+++ src/core/Core/Debug/DebugUIWorld.cpp
@@ -91,26 +91,26 @@ void DebugUI::show_world_inspector() {
     }
     ImGui::TextDisabled("Session ownership is independent of CTL controller enablement.");
     if (m_context.scenario_engine != nullptr &&
-      m_context.scenario_engine->current_character_trigger_proxy().has_value()) {
+        m_context.scenario_engine->current_character_trigger_proxy().has_value()) {
       const CurrentCharacterTriggerProxy& proxy{
-        m_context.scenario_engine->current_character_trigger_proxy().value()};
+          m_context.scenario_engine->current_character_trigger_proxy().value()};
       ImGui::Text("Trigger proxy: %s | owner character %d, world %u | generation %llu",
-        proxy.registered ? "registered" : "unregistered",
-        proxy.owner.character_id,
-        proxy.owner.world_scene_id,
-        static_cast<unsigned long long>(proxy.generation));
-        ImGui::Text("Proxy contact readiness: %s", proxy.contact_ready ? "ready" : "armed");
+          proxy.registered ? "registered" : "unregistered",
+          proxy.owner.character_id,
+          proxy.owner.world_scene_id,
+          static_cast<unsigned long long>(proxy.generation));
+      ImGui::Text("Proxy contact readiness: %s", proxy.contact_ready ? "ready" : "armed");
       ImGui::Text("Proxy XYZ: %.3f, %.3f, %.3f | radius %.3f | heading %.3f deg",
-        static_cast<double>(proxy.position.x),
-        static_cast<double>(proxy.position.y),
-        static_cast<double>(proxy.position.z),
-        static_cast<double>(proxy.radius),
-        static_cast<double>(proxy.heading_degrees));
-        ImGui::Text("Proxy overlapping zone contacts: %zu", proxy.overlapping_zone_count);
+          static_cast<double>(proxy.position.x),
+          static_cast<double>(proxy.position.y),
+          static_cast<double>(proxy.position.z),
+          static_cast<double>(proxy.radius),
+          static_cast<double>(proxy.heading_degrees));
+      ImGui::Text("Proxy overlapping zone contacts: %zu", proxy.overlapping_zone_count);
       ImGui::Text("Proxy synchronization: %s%s%s",
-        proxy.synchronization_suspended ? "frozen" : "ordinary actor update",
-        proxy.suspension_reason.empty() ? "" : " - ",
-        proxy.suspension_reason.c_str());
+          proxy.synchronization_suspended ? "frozen" : "ordinary actor update",
+          proxy.suspension_reason.empty() ? "" : " - ",
+          proxy.suspension_reason.c_str());
     } else {
       ImGui::TextUnformatted("Trigger proxy: not registered");
     }
@@ -419,8 +419,7 @@ void DebugUI::show_world_inspector() {
             character.ctl_move_name.c_str());
         ImGui::Text("State: %d, animation key: %s",
             static_cast<int>(character.ctl_state_id.value_or(0)),
-            character.ctl_animation_key.empty() ? "<none>"
-                                                : character.ctl_animation_key.c_str());
+            character.ctl_animation_key.empty() ? "<none>" : character.ctl_animation_key.c_str());
         ImGui::Text("Phase: %.3f -> %.3f (end %.3f)",
             static_cast<double>(character.ctl_previous_progress),
             static_cast<double>(character.ctl_current_progress),
@@ -781,10 +780,10 @@ void DebugUI::show_sprite_instances_tab(
     ScenarioRuntime& runtime, SceneDebugView* const scene_view, const float delta_time) {
   Sprite::SpritePool& pool{runtime.sprite_pool()};
 
-  ImGui::Text("Pool: %lu live / %lu capacity / %lu attached",
-      static_cast<unsigned long>(pool.live_count()),
-      static_cast<unsigned long>(pool.capacity()),
-      static_cast<unsigned long>(pool.attached_count()));
+  ImGui::Text("Pool: %zu live / %zu capacity / %zu attached",
+      pool.live_count(),
+      pool.capacity(),
+      pool.attached_count());
 
   if (ImGui::BeginChild("##SpriteInstances", ImVec2(0.0F, 120.0F), ImGuiChildFlags_Borders)) {
     for (auto head{pool.render_list_head()}; head.has_value();
